# Utilize ext data location to reduce qd matmul memory usage