@skottmckay 2024-07-23
What scenario requires this vs. say optimizing the model to level 2 offline for now? We should be really clear on whether it's required before checking in a hack.

It feels a little intrusive so I wonder if we're maybe solving the wrong problem. 

This comes up semi-regularly, so the need to hack around the current setup may be a good forcing function for making a change to when we convert from a model storage specific data type (TensorProto) to ORT data types (OrtValue). i.e. it may be better to create OrtValue instances for initializers during model loading. Doing so potentially enables additional optimizations like doing an in-place transpose so we only need a temporary buffer for the transpose before storing the result in the original memory allocation.  

Not trivial though. To change from TensorProto to OrtValue during model loading would mean all the optimizers need to be updated. We could maybe use an adapter class in the optimizers and other ORT code that expects protobuf data types to pretend an OrtValue is a TensorProto so we can migrate code gradually. 

And if we're going to break the connection on protobuf data types during model load the changes required also extend to any other usage of protobuf types in the code like AttributeProto in the onnxruntime::Node class.