@fajin-corp 2024-07-24
This is going to be a good discussion.

These changes are for QDQ models. If every time when a user wants to run a large QDQ model in ORT, he/she needs to offline optimize the model first, that will probably drive the user away from ORT.

I second "it may be better to create OrtValue instances for initializers during model loading". Currently I don't have good idea of how to do it. TensorProto is final. How to make code to accept both TensorProto and OrtValue is not clear to me. Would greatly appreciate if you can share ideas. We can start from TensorProto, which is the real headache. Then extend the approach to other protobuf types.

---
In reply to: [2244451018](https://github.com/microsoft/onnxruntime/pull/21451#issuecomment-2244451018) [](http://example.com/codeflow?ancestors=2244451018)