@yufenglee 2024-07-29
> l 2 offline for now? We should be really clear on whether it's required before checking in a hack.

@skottmckay, this happens in the fusion of DQ + MatMul to MatMulNBits and this is for the case that users need a model with only standard onnx operators.

For the alternative proposal (converting TensorProto to OrtValue_, it need a total re-design of the workflow. In addition to what you describe, the Graph is also built on top of protobuf, like Graph.Resolve. We need to re-design the Graph on top of OrtValue too.