@fajin-corp 2024-07-30
/azp run Windows GPU CI Pipeline